renamed AdminPages to AdminPage
